@@ -14,7 +14,7 @@
     "vendor"
   ],
   "types": "bson.d.ts",
-  "version": "6.3.0",
+  "version": "6.4.0",
   "author": {
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
@@ -27,39 +27,39 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.38.4",
-    "@octokit/core": "^4.2.4",
+    "@microsoft/api-extractor": "^7.40.5",
+    "@octokit/core": "^5.1.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-typescript": "^11.1.5",
+    "@rollup/plugin-typescript": "^11.1.6",
     "@types/chai": "^4.3.11",
     "@types/mocha": "^10.0.6",
-    "@types/node": "^18.19.2",
-    "@types/sinon": "^10.0.20",
+    "@types/node": "^20.11.19",
+    "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^3.2.12",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
     "benchmark": "^2.1.4",
     "chai": "^4.3.10",
     "chalk": "^5.3.0",
     "dbx-js-tools": "github:mongodb-js/dbx-js-tools",
-    "eslint": "^8.55.0",
-    "eslint-config-prettier": "^8.10.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-no-bigint-usage": "file:etc/eslint/no-bigint-usage",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-tsdoc": "^0.2.17",
-    "magic-string": "^0.30.5",
-    "mocha": "10.2.0",
+    "magic-string": "^0.30.7",
+    "mocha": "10.3.0",
     "node-fetch": "^3.3.2",
     "nyc": "^15.1.0",
-    "prettier": "^2.8.8",
-    "rollup": "^3.29.4",
-    "sinon": "^15.2.0",
+    "prettier": "^3.2.5",
+    "rollup": "^4.12.0",
+    "sinon": "^17.0.1",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
     "standard-version": "^9.5.0",
     "tar": "^6.2.0",
-    "ts-node": "^10.9.1",
-    "tsd": "^0.28.1",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.30.5",
     "typescript": "^5.0.4",
     "typescript-cached-transpile": "0.0.6",
     "uuid": "^9.0.1"
