@@ -1,6 +1,6 @@
 {
   "name": "mongodb",
-  "version": "6.3.0",
+  "version": "6.4.0",
   "description": "The official MongoDB driver for Node.js",
   "main": "lib/index.js",
   "files": [
@@ -26,7 +26,7 @@
   },
   "dependencies": {
     "@mongodb-js/saslprep": "^1.1.0",
-    "bson": "^6.2.0",
+    "bson": "^6.4.0",
     "mongodb-connection-string-url": "^3.0.0"
   },
   "peerDependencies": {
@@ -62,53 +62,54 @@
     }
   },
   "devDependencies": {
+    "@aws-sdk/credential-providers": "^3.515.0",
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.36.4",
+    "@microsoft/api-extractor": "^7.40.6",
     "@microsoft/tsdoc-config": "^0.16.2",
-    "@mongodb-js/zstd": "^1.1.0",
-    "@octokit/core": "^4.2.4",
-    "@types/chai": "^4.3.5",
-    "@types/chai-subset": "^1.3.3",
-    "@types/express": "^4.17.17",
-    "@types/kerberos": "^1.1.2",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^20.5.9",
-    "@types/saslprep": "^1.0.1",
-    "@types/semver": "^7.5.0",
-    "@types/sinon": "^10.0.16",
-    "@types/sinon-chai": "^3.2.9",
-    "@types/whatwg-url": "^11.0.0",
+    "@mongodb-js/zstd": "^1.2.0",
+    "@octokit/core": "^5.1.0",
+    "@types/chai": "^4.3.11",
+    "@types/chai-subset": "^1.3.5",
+    "@types/express": "^4.17.21",
+    "@types/kerberos": "^1.1.5",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.11.20",
+    "@types/saslprep": "^1.0.3",
+    "@types/semver": "^7.5.7",
+    "@types/sinon": "^10.0.20",
+    "@types/sinon-chai": "^3.2.12",
+    "@types/whatwg-url": "^11.0.4",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "chai": "^4.3.7",
+    "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^8.48.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "eslint-plugin-unused-imports": "^2.0.0",
     "express": "^4.18.2",
-    "gcp-metadata": "^5.2.0",
+    "gcp-metadata": "^5.3.0",
     "js-yaml": "^4.1.0",
-    "mocha": "^10.2.0",
+    "mocha": "^10.3.0",
     "mocha-sinon": "^2.1.2",
     "mongodb-client-encryption": "^6.0.0",
-    "mongodb-legacy": "^6.0.0",
+    "mongodb-legacy": "^6.0.1",
     "nyc": "^15.1.0",
     "prettier": "^2.8.8",
-    "semver": "^7.5.4",
+    "semver": "^7.6.0",
     "sinon": "^15.2.0",
     "sinon-chai": "^3.7.0",
     "snappy": "^7.2.2",
-    "socks": "^2.7.1",
+    "socks": "^2.8.1",
     "source-map-support": "^0.5.21",
-    "ts-node": "^10.9.1",
-    "tsd": "^0.28.1",
-    "typescript": "^5.0.4",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.30.6",
+    "typescript": "5.0",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.3.1",
     "yargs": "^17.7.2"
