@@ -1,26 +1,22 @@
-const MenuModel = require("../model/Menu");
+import MenuModel from "../model/MenuModel.js";
 
-class MenuController {
-  // [POST]
-  addMenu(req, res) {
-    MenuModel.create(req.body)
-      .then((menu) => {
-        res.json({ success: true, message: "add menu success", data: menu });
-      })
-      .catch((error) => {
-        res.json({ success: false, message: "add menu failed", error: error });
-      });
+export const getAllMenu = async (req, res) => {
+  try {
+    const menu = await MenuModel.find();
+    res
+      .status(200)
+      .json({ success: true, message: "get menu success", data: menu });
+  } catch (error) {
+    res.status(500).json({ success: false, message: "get menu failed" });
   }
-  // [GET]
-  getMenu(req, res) {
-    MenuModel.find()
-      .then((menu) => {
-        res.json({ success: true, message: "get menu success", data: menu });
-      })
-      .catch((error) => {
-        res.json({ success: false, message: "get menu failed", error: error });
-      });
-  }
-}
+};
 
-module.exports = new MenuController();
+export const getSingleMenu = async (req, res) => {
+  const id = req.params.id;
+  try {
+    const menu = await MenuModel.findById(id);
+    res.status(200).json({ success: true, message: "get success", data: menu });
+  } catch (error) {
+    res.status(500).json({ success: false, message: "get failed" });
+  }
+};
